Move job commit tests into new class
